tdf#156699 a11y: Set a11y name for SvxCharView
Set the character name as a11y name for the `SvXCharView`.

Also add a check for the string to be non-empty in
`SvxCharView::GetDecimalValueAndCharName`, since
`SfxCharmapContainer::updateRecentCharControl` sets the
text to an empty string, and calling
`OUString::iterateCodePoint` on that empts string
would trigger an assert now that
`SvxCharView::SetText` tries to retrieve the character
name right away.

This makes NVDA on Windows announce the focused
character by its name in the Special Characters dialog
and the Special Characters toolbar group button.

Orca with the gtk3 VCL plugin now announces the
focused character in the Special Characters dialog,
but not yet in the toolbar group button.

Change-Id: I79c8e0c2de6e817a467a405dc7d9e103cba07dfe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155757
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>