tdf#156728 sw: layout: invalidate fly on header/footer size
On page 30 there are at-page anchored shapes Shape22, Shape23 and
Shape25.

Shape22 and Shape23 are positioned relative to the bottom of the page
text area, and that is working; Shape25 is positioned with an offset to
the top of the page text area, and that doesn't work.

The problem is that the shapes are positioned when the header on the
page hasn't been formatted yet, and once the header is formatted, the
position of the Shape25 is not invalidated, so it's a header-height too
far up.

(regression from commit 425a252c3cc5e5a79a533965026dd4af6b8df739)

Change-Id: Ic907d98b1bcd4bb544fd1af9bf93ebc7b75bf362
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155762
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>