Template: tdf#86483: Add Japanese template to the Localization category
As Kevin Suo noted in tdf#86483, CJK and CTL use different formatting than
the US and Europe. Japanese is no exception, and must be formatted according
to Japanese typesetting rules[1][2] before the document is created. However,
many Japanese language users use LibreOffice without knowing this, resulting
in strange Japanese documents. This template sets up the necessary Japanese
formatting to reduce this.

Originally, the template was supposed to be provided by the language pack.
However, since a localization category was created, we thought it would be
a good idea to provide the template first, as we did with Chinese, so we
added it.

This template has the following settings And the template is inspired by
the OpenOffice.org Japanese Environment Improvement Extension Template[3]
and Kevin Suo's Chinese template, which we have previously attempted to do
the same.

* Western and Asian languages are in the same font.
* Paragraphs have been set to be justified.
* Japanese paragraphs are originally indented by one character. However,
  since some documents do not use it, it is not set. Instead, the original
  Japanese paragraphs are set to First line indent style.
* Italics has been removed from the style. Japanese does not use italics.

[1] Requirements for Japanese Text Layout:
    <https://www.w3.org/TR/jlreq/>
[2] Japanese Industrial Standards - JIS X 4051:2004(Japanese):
    <https://kikakurui.com/x4/X4051-2004-02.html>
[3] OpenOffice.org Japanese Environment Improvement Extension:
    <https://ja.osdn.net/projects/openoffice-docj/releases/40804#note>

Change-Id: Ieaf57c8e018d479e7922fbd9d616b9a6c271b1ba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155764
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>