tdf#156561 a11y: Don't ignore child events in AccessibleListBox
Don't ignore the `WindowShow` and `WindowHide` child events
in `AccessibleListBox::ProcessWindowChildEvent`, but let the
base class implemenation handle them just like all other
child events.

This causes them to be forwarded as `AccessibleEventId::CHILD`
events to the a11y layer, which e.g. triggers registering
a11y event listeners for newly created/shown child objects
in winaccessibility.

This makes the announcement of sub menu entries in Calc's auto
filter work even after reverting

        commit dc0706cabfe39ddb6ea23d60ccfb756f2b9e6efb
        Author: Michael Weghorn <m.weghorn@posteo.de>
        Date:   Wed Mar 15 17:00:27 2023 +0100

            tdf#140762 tdf#152671 Make dock win visible before showing popup

, which will be done in a follow-up commit.

`AccessibleListBox` is the a11y class used for
tree views in the non-gtk case (i.e. `SalInstanceTreeView`).
A tree view is passed as the parent of the sub menus in
the autofilter popup (s. `ScCheckListMenuControl::addSubMenuItem`),
but due to ignoring the events, the submenu would not be
announced by NVDA before the above-mentioned commit.

The events were ignored since

    commit 4c01898e02d7b80f70f19de12f2598644fc421e3
    Author: Vladimir Glazounov <vg@openoffice.org>
    Date:   Mon Jan 28 13:14:03 2008 +0000

        INTEGRATION: CWS tbe32 (1.2.16); FILE MERGED
        2007/11/05 15:14:07 tbe 1.2.16.1: #i70908# [A11y] OOo crashes in Customize dialog - Add when large icons and A11y ONcommit 4c01898e02d7b80f70f19de12f2598644fc421e3

to fix a crash in the customization dialog,
but that problem described in [1] is no longer
reproducible for me with the change here
in place.

Comment 8 in the ticket mentions:

> The problem was the accessible child event, which was sent for the
> help text window. This event triggered some callbacks which caused
> the dead lock. As the help text window is not exposed as accessible
> child of the tree list box, I removed the accessible child event.

The fact that extra children from the `vcl::Window`
hierarchy (besides the entries in the `SvTreeListBox`
are currently not exposed via
`AccessibleListBox::getAccessibleChildCount` and
`AccessibleListBox::getAccessibleChild` is still
true, but not a problem in this context.
(If necessary, handling for those should be easy to
add by forwarding to the corresponding
`VCLXAccessibleComponent` methods, and adapting
the child index used as needed.)

[1] https://bz.apache.org/ooo/show_bug.cgi?id=70908
[2] https://bz.apache.org/ooo/show_bug.cgi?id=70908#c8

Change-Id: If374032387babf41b28067d5df54d5f6ce682c48
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155801
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>