tdf#151059 don't use Ctrl+Alt+3 etc by default for gallery
because on some keyboard layouts ctrl+alt is basically
equivalent to altgr and removes the possibility to enter
specific letters.

use Alt+3 etc because that 1-9 sequence seems free and
unproblematic vs other combinations.

If that doesn't work out then maybe we don't need one per sidebar entry
and just have a single entry to jump to the sidebar toolbar and a follow
up of press of 1-9 to activate a specific deck.

Change-Id: I676d03c7e1c7bcb3b6c08e6d30aa23cf61838b38
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155831
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>