svx: remove theme related parameters for Color and FillColor
For SvxColorItem, XFillColorItem, XLineColorItem,... we set the
ComplexColor via the JSON additional UNO Command argument, so we
don't need the parameters to set the theme color and LumOff and
LumMod color modifiers for the UNO Command. Actually this has
broken the setting the colors in certain situations, so we must
remove this.

Change-Id: Ic28aa1e43559d3627703d93c7c1bd9310a149986
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155864
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>