crashtesting: parallel calc assert with Lookup
seen in forum-mso-en4-55192.xlsx with a singlecell as the result vector
which gets autoextended outside the range that the pre-parallel calc
took into consideration on set up. For simplicity if there is a result
vector assume that if it doesn't match the size of the search vector
that we can't parallelize this.

Change-Id: Ic67e1b5b35964760ac5b1608cd516a69e08d0540
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155862
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>