tdf#156610 tdf#139633 SdNavigator: Fix object not selected
for entry that context menu is shown for, GtkInstance only

Commit fdb504add8e21c137a37fde48fedb93638e8ee97 introduced a hack to
make inplace editing work for the SalInstance of the treeview that,
for Gtk3, breaks object selection before the entry popup menu is
displayed which causes the menu item Rename... not to always work as
expected. Sal doesn't notice this because the mouse pressed signal is
seen when there is a popup menu handler set. Gtk3 does not fire the
mouse pressed signal when a popup menu handler set. This patch makes
object selection always happen before the popup menu is shown.

Change-Id: I562ddf99bd15a48631170a7d70821021a4139295
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155867
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>