tdf#156351 sw floattable: fix missing bottom border in master table
The problem was that the bugdoc has a split floating table, and the
bottom of the table on the first page was missing its border. The
borders on the second page were correct.

This happens because the cell itself has no bottom border and the layout
did not extra effort to mirror the top table margin at the buttom, like
Word does.

Fix the problem similar to commit
53798fef2cc0b5b0b9706081a4af5ceca964a41b (Related: tdf#156351 sw
floattable: fix missing top border in follow table, 2023-08-18), so we
add a bottom border for master tables at a layout level, similar how a
top border was already added for follow tables.

Given that this kind of worked already in the past (just top borders,
just the 1 row case), do this unconditionally; but if needed this could
be limited to the "Word table cell" case.

Change-Id: Iafdcd90226fdc425be597d36ad97fb69dca5a89a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155884
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins