tdf#137523 sw: layout: fix text below tables in footer
In the footer, the text frames below a table have a very small height
and are effectively invisible.

This is because the JoinLock in SwLayAction::FormatLayoutTab() causes
SwTabFrame::MakeAll() to return early, which is called for every cell
frame in the table, and which would move the table up the page in
MakePos(), by calling SwHeadFootFrame::Format().

(regression from commit cc5916cd314a27b0cc99560ab887480026630a95)

Change-Id: I4add9fdd73ad3756000969af20ebf758025a0ada
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155902
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>