tdf#155510 ODT import: ignore props of covered table cells, unless in comp mode
The problem was that the A5 cell on page 2 used to have a solid
background, but this was lost after commit
1001dbaef4dec2b51c25ed8343bab6910f1219e1 (ODT import: handle style name
of covered cells, 2021-03-19).

Covered cell formatting was interesting for cell borders, in Word compat
mode the covered cell can have its own border style, so e.g. there is
left border on a vertically merged cell till its center, but not after
its center.

Fix the problem by limiting the import of covered cell formatting to the
case when the TABLE_ROW_KEEP compat flag is on, to keep both native ODT
and ODT-from-Word documents happy.

If there is a need, some smarter way of merging properties could be
implemented in the future.

Change-Id: Icc1e6f48d9f5a0379d9a9052a1f5b9ae3f294477
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155957
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>