tdf#77760 sw floattable: add support for footnotes, layout
Once the footnote was inserted to the document model, the layout was
also missing (footnote portion, footnote frame).

One problem is that the footnote portion was empty instead of the
footnote counter, because SwTextFormatter::NewFootnotePortion() returned
early in the fly case.

The other problem was the missing footnote frame, but once the footnote
portion was created correctly, the matching footnote frame was also
created.

File filters (ODT, DOCX) still needs more work.

Change-Id: If71c44962f604f23fdcfccfe80b0d97787fd99d4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155961
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins