tdf#137523 sw: layout: fix text below tables in footer differently
This fix is better than commit 027f8328eef1e149b6c99b478ed5df870291dc2d
because it turns out the last text frame had a height of 219 in 6.1.

One of the following commits would break this, so a different fix is
needed.

In SwFrameNotify::ImplDestroy(), if the frame changes its position the
next frame's position is invalidated; if the next frame is undersized
then also invalidate its size so that it will be formatted again, with
more space.

Change-Id: I8e6a3fe3127ebfa2246c440d2a3455b217476475
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155986
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>