tdf#154775 sw: layout: avoid breaking this with following commits
The problem with commit 1b5bc2ae2f19a190bf588a5a26c4d125c3960991 is that
the fix only works in case the fly is invalid at the time when
SwTabFrame::CalcFlyOffsets() is called; if it is valid (which will be
the case with following commits, for non-obvious reasons) then the fix
does nothing, causing testTablePrintAreaLeft to fail.

In Word 2013, this fly causes the table to shift down, despite its
HoriOrient being LEFT; so adapt SwTabFrame::CalcFlyOffsets() to also
shift down in this case.

Likely it would be best not to check HoriOrient at all for ShiftDown
case, but needs checking first what Word does in other cases.

Change-Id: I465a55bbf1a783abce8b8d6d65099bd9df5695f8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155987
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>