tdf#114441 - Convert use of sal_uLong to better integer types
nUpdateInterval (sal_uLong --> sal_uInt64) because it is passed as
argument to Timer::SetTimeout function that takes a sal_uInt64

func CalcUpdateInterval (sal_uLong --> sal_uInt64 return type) because
this function is only being called once, and in that call, the
return value is passed to a function which takes sal_uInt64 type
int. line 412
(https://opengrok.libreoffice.org/xref/core/sc/source/ui/view/select.cxx?r=b6b26421#412)
two lines after which it is passed to a function ...

and the function changed in vcl/source/window/seleng.cxx is the one to
which the value of nUpdateInterval is being passed as explained in
the lines above.

Change-Id: I73441c2d3520a222ddb39b8c0c205b4dd61d4751
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156015
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>