sw floattable: don't split if anchored inside a footnote
See
<https://gerrit.libreoffice.org/c/core/+/156025/1#message-94d8cdd4d59a4b991186e5780d41f213105a533d>,
if we import an (inline) table inside a footnote from DOCX, we currently
put that into an as-char anchored fly. In case the user would later
change the anchor type to to-para, then we would crash in
SwFlowFrame::GetPrevFrameForUpperSpaceCalc_().

Avoid all this trouble by simply not allowing split floating tables in
footnotes, seeing that the Word UI is also disabled for this case.

Change-Id: I3cf9461beb291084e1cced5c66ca14e6b6d55126
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156058
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156644