tdf#77760 sw floattable: add support for footnotes, DOC import
This is similar to commit 178421a6c719dac9c16f220b76292fec16a53f60
(tdf#77760 sw floattable: add support for footnotes, DOCX import,
2023-08-24), the problematic part was to reject everything that is not
in the body text, relax that to allow insertion into split flys.

Do an early check to see if we'll insert into the fly/header/footer
section, because otherwise it would be pointless to call
SwNode::GetFlyFormat(), which can be expensive in case we don't have a
layout yet.

The DOC export, the RTF import and the RTF export was working already,
so filters are mostly covered with this.

Change-Id: I59c69fac0692c6656c054e32503ec0cbc2fd11e0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156083
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins